Move Loop and LazyMap into their own files
