+++ blueeyes/src/main/scala/quasar/blueeyes/Loop.scala
@@ -16,8 +16,6 @@
 
 package quasar.blueeyes
 
-import java.util.concurrent.ConcurrentHashMap
-
 import scala.annotation.tailrec
 
 /**
@@ -40,22 +38,4 @@ object Loop {
 
     loop(0)
   }
-}
-
-final class LazyMap[A, B, C](source: Map[A, B], f: B => C) extends Map[A, C] {
-  private val m = new ConcurrentHashMap[A, C]()
-
-  def iterator: Iterator[(A, C)] = source.keysIterator map (a => a -> apply(a))
-
-  def get(a: A): Option[C] = m get a match {
-    case null =>
-      source get a map { b =>
-        val c = f(b)
-        m.putIfAbsent(a, c)
-        c
-      }
-    case x => Some(x)
-  }
-  def +[C1 >: C](kv: (A, C1)): Map[A, C1] = iterator.toMap + kv
-  def -(a: A): Map[A, C]                  = iterator.toMap - a
 }
